MB-43764 [BP MB-42623+44147 to 6.6.2] Fix descending sort edge case
(Backport omits unrealted code cleanups and comment changes.
Original MB-43764 patch was buggy, replaced by MB-44147 patch.)

Scans of partitioned indexes that request the primary key
plus all GSI keys in the same order they are defined in the
index, where at least one of those keys is indexed in
descending order, were returning unsorted data because
GsiClient code analyzeProjection() (in
queryport/client/scatter.go) was making an incorrect
assumption that if sorted data is required, there will
always be a projection list, but in this case that is not
true, so it was not setting up the scan request's projDesc
slice for this case.

In the scan case being fixed, when there is no projection list,
we just need to initialize the RequestBroker.projDesc slice as
a reference to the IndexDefn.Desc slice rather than creating a
new slice and populating its entries based on key part position
lookups.

Change-Id: I55050368ec3fd7bc551c09abaee5838ff7bb686c