MB-42743 Pass filtered version of indexInstList for snapshot generation
When there are large number of indexes spread across buckets, iterating
through all the indexes to indetify the index belonging to the keyspace
of interest is a costly operation - as this iteration has to be done
in the snapshotting code path.

Hence, creating a filtered version of index instance ID's according
to the stream and keyspace. This makes sure that only the instances
belonging to the stream and keyspace of interest will be iterated
while generating snapshots

Change-Id: I5360c40f05c90d41832ac1ca84dd8f44f435f3b6