MB-44269 Use tokens returned by metakv.ListAllChildren
Fixed places in GSI code that were (directly or indirectly) calling
ns_server metakv.ListAllChildren() and then metakv.Get() on the
path of each child to retrieve the child token. The child tokens
are already returned by the first call in KVEntry.Value, so we do
not need to make subsequent calls to retrieve each token one by one.

Change-Id: Ie67c6b6954290582b220d410786cc64c23fea0a9