MB-44559 Increase the size of builder notifyCh to avoid deadlocks
Indexer only allows building indexes of 10 collections at a time.
When there are 10000 build requests with 10 indexes per collection,
100 indexes will be built while lifecycle manager will try to write
remaining 9900 index definitions to notifyCh. Also, for those instances
that are being built, lifecycleMgr::handleTopologyChange would write
the instance state change into notifyCh. This means that there are ~10000
notifications being written into notifyCh while the capacity of the channel
is also 10000.

This is leading to a deadlock as lifecycle manager tries to write to an
already full notifyCh and gets blocked and the consumer of notifyCh is
waiting for lifecycle manager to consume a message to build index

Change-Id: Ia089ac64759348ce0215c12986014bf8db348ae0