MB-43967 Part 3: getIndexStatus ETag for full results to caller
Part 3 is the final part of the getIndexStatus ETag feature.
It implements computation and use of ETags to avoid resending
unchanged full sets of index status info back to the caller
(usually ns_server).

It also adds a disk caching performance improvement that skips
rewriting unchanged results back to the disk cache.

Change-Id: I980e5e8c3fb80f83e9c322b95f376825fa64adb6