MB-44442 Reduce the default size of VbucketWorker encodeBuf
Currently, the defauly size of VbucketWorker encodeBuf is
set at 1MB. With collections, it is possible that the number
of streams managed by projector can significantly increase.

In this issue, 1920 VBucketWorkers were active at the time
of cbcollect. This means projector requires a minimum of
1.92G of memory just to manage empty streams. In this, patch
we reduce the default size of the encodeBuf size to 2KB per
worker. As the buffer gets resized when a larger sized mutation
is proceessed, there should not be any functional impact because
of this change

Change-Id: Ib36ec45c5983584b92c378eca686735d4724dfa1