MB-44764 Re-order config check and checkInterval() evaluations
The checkInterval method calls isInitial() which inturn iterates
over entire list of instances to see if there is any instance
belonging to the streamId and keyspaceId on which flush was done.

With 10000 indexes on a node, iterating over all 10000 indexes
for every flushDone is a costly call. This can be avoided when
writer tuning is not enabled. Hence, re-ordering the checks
so that we skip the iteration of 10000 indexes

Change-Id: I0b96960dbe8d1375fdf62d409d97d84a82027384