MB-43427 Fix negative client pending item stats
Client pending item stats are initialized to math.MaxInt64 to
distinguish between when the client has not yet received the
stats versus actual zeros. For logging, these stats values are
summed across partitions, resulting in a negative value if an
index is partitioned.

The fix is to return math.MaxInt64 as the sum if any partition
has this value for its stat, because the sum of anything plus
<unknown> is garbage. Now the logs will show the initial value
of 9223372036854775807 regardless of how many partitions exist
instead of negative numbers.

Change-Id: I596f7d5e132e16277c04429ef1a23a75b08ed7ad