MB-45160 Fix ETag interaction with pre-Cheshire-Cat nodes
Pre-Cheshire-Cat nodes do not compute ETags, thus LocalIndexMetadata
(LIM) returned from them do not have the ETag or ETagExpiry field set
as they would if they came from a CC node. In computing the global
ETag getIndexStatus returns to caller, zero ETags in LIMs from down-
level nodes triggered a sanity check error in request_handler.go
setETagGetIndexStatus().

The fix is to compute the ETag and ETagExpiry locally on the CC
getIndexStatus handling node (*before* mem or disk caching them). An
alternate fix would be to return a zero ETag from getIndexStatus(),
however qualitative analysis says this alternate would perform
worse on average than the chosen fix in a mixed-level cluster:

  1. small = cost of indexer marshal LIM and compute ETag
  2. small = cost of marshal and send LIM from indexer to ns_server
     (small because it is local, not network, though it is via HTTP)
  3. large = cost of ns_server unmarshal and process LIM

For every LIM, the chosen fix pays cost 1 and almost never costs
2 or 3, though occasionally it will pay all three costs. In the
alternate fix, every LIM pays both costs 2 and 3 100% of the time,
so on average this is worse. Moreover, in the alternate fix a single
down-level node would require sending results from *all* nodes back
to ns_server, so it forces the costs 2 and 3 multiplied by the number
of nodes instead of just for the down-level node. Note also that all
these costs are paid by the same node, just by different services, as
ns_server calls /getIndexStatus on the same node it is running on. It
does no good to avoid a small cost (1) in indexer just to generate a
different small cost (2) in indexer plus a large cost (3) in ns_server.

Change-Id: I72e9d7bb8353c4992497b07cfdd669fc3db6aa76