MB-45210 : Fix indexes getting dropped during upgrade from 6.5.1
Also fixing the error seen due to hitting the scopes endpoint when the
cluster is not upgraded to 7.0 and indexer is upgraded.

Getting empty collection ID from cinfo as cluster is not upgraded.
Collection ID restored in Defintion is filled with defaults and hence the
indexes are not being restored.

Change-Id: I7d09de74bc72ba62caba85ea750f8ea76bd91981