MB-45230 fix checking for pending writer before storage snapshot
Golang doesn't provide any instruction ordering guarantee as observed
from an external goroutine, if no sychronization mechanism/channel
communication has been used.

This could result in the existing pending writer check(waitPersist),
based on the value of atomic qCount, to go wrong if the storage insert/delete
function call has not returned before the qCount is set by command
workers.

This patch addresses the problem by setting the qCount in a defer call,
which is guaranteed by the golang runtime to execute after the function
body has finished executing.

Defer function call has an overhead of about 35ns in Go 1.13. This has
been optimized to be about 6ns in Go 1.14. Given all the processing
which needs to happen for a mutation(GSI layer and storage), the
pipeline should be able to absorb this extra cost.

e.g. if indexer can process 1M mutations/sec on a 50 core box, that's
about 20k mutations/sec on a single core. This translates to roughly
50us for a single mutation. Adding a 35ns overhead to this looks fine.

The other option considered was to make the setting of qCount
conditional on the return value of the storage insert/delete function.
This approach should also work but is still not backed by any runtime
guarantees. If there is any perf regression, this option can be
considered.

Change-Id: I9f9e23f6ab87ad44275e9a95ebabc3fd60f3b623