MB-44728 Part 1: Fix rebalance create, drop cases of b/s/c not found
This part prevents rebalance from failing during the index metadata
create (destination) and index drop (source) steps of rebalance due
to the bucket, scope, or collection referenced by the transfer token
having been dropped between token creation and token processing.

The fix is for rebalancer to check whether the error string is the
same as one of those from ErrBucketNotFound, ErrScopeNotFound,
or ErrCollectionNotFound and skip ahead (or in the source case just
continue on to) TransferTokenCommit state. This will abort the move
of the index without failing the rebalance.

Note: most of Part 1 changes are actually comments added. I had to
spend a lot of time figuring out the details of how rebalance works
so I added comment headers to most of the functions and some of the
class variables I was looking at. The fix is in rebalancer.go
processTokenAsDest() and dropIndexWhenIdle(), plus the new helper
function isMissingBSC().

Change-Id: I37df7c6cad12b0bc3aa294f19b77aed8442e9e22