MB-44764 Use stateless worker pool for snapshot generation
Instead of spawning 10000 go-routines when there are 10000
indexes, this patch will spawn go-routines equivalent to
the config "numSnapshotWorkers" (say 1000) for example
and these 1000 snapshot workers will generate snapshots for
all 10000 indexes. So, 1 go-routine will generate snapshots
for 10 indexes.

This will reduce runtime scheduling overhead on golang and
speed-up snapshotting. As the load balancing is static in
nature, if any index takes long time to flush mutations,
the snapshots for ther indexes that are in the batch waiting
to be processed after this index will be delayed. This is
the only downside of this patch. In such a case, the
"numSnapshotWorkers" config variable can be adjusted to
suite the number of instances in the keyspace

Change-Id: I62767db4f2734b523cac84c7aefd602d1e39d7f1