MB-45400 Fix getStatsFromCache() self-deadlock on statsMutex
Due to a superfluous "defer" keyword, request_handler.go
getStatsFromCache() was holding statsMutex read-locked when calling
memCacheStats(), which tries to write-lock the same mutex. The mutex
should be released immediately after accessing statsCache[].

Refactored the code (and similar but non-buggy code in
getLocalIndexMetadataFromCache()) to immediately release the mutex
and also eliminate the need for two separate RUnlock() calls.

Change-Id: I74ebecc48b1c5fc3ebfbc301e7f0481ae82bae5f