MB-45329 destroy index snapshot for deleted index
Storage Manager can notify a new snapshot to scan coordinator.
This is read asynchronously in listenSnapshot. It is possible for index
to get deleted and updateLastSnapshotMap to clear the lastSnapshot
before listenSnapshot gets a chance to process the notification.
When listenSnapshot processes the notification, it will find that there
is no entry in lastSnapshot and ignore this new snapshot. This snapshot
will never get destroyed.

Change-Id: I5e775dc92fd5c46dd9b4be5b39c1eb732fb9ac1f