MB-44728 Part 2: Fix rebalance build cases of b/s/c not found
This part prevents rebalance from failing during the index build
step of rebalance due to the bucket, scope, collection, or index
referenced by the transfer token having been dropped between token
creation and token processing.

The fix requires returning detailed per-instId error information
in an error map from Indexer handleBuildIndex() (lightly edited by
LifecycleMgr buildIndexesLifecycleMgr() on the way back up) and
received and processed by Rebalancer buildAcceptedIndexes(). In the
rebalance case only, this chain now returns a marshaled JSON string
of a map[InstId]string where the string is the error message, instead
of the former single generic error string if any build failed. In
cases where the instId could not be found, the map key is actually the
originally submitted defnId, which is detected because in that case
the associated error string will be set to the special error
common.ErrIndexNotFoundRebal.Error() which is created for this purpose
and used nowhere else.

The following rebalance build errors are now handled by moving the
associated transfer token(s) forward to TransferTokenCommit state
instead of failing the rebalance:
  1. instId ErrBucketNotFound
  2. instId ErrScopeNotFound
  3. instId ErrCollectionNotFound
  4. defnId ErrIndexNotFoundRebal
1-3 update the instId's transfer token. 4 updates all TTs that are
part of defnId.

Change-Id: I0de31ba3eae7d558cf2f9cc79fb153055fc5ad8f