MB-45228 Relax GC when projector RSS is consistently <1% of memTotal
We pesimistically adjust the GOGC percent when projector RSS
is consistently <1% of memTotal. Based on the projector RSS,
the GC percent is relaxed upto 400%. We call this a pesimistic
approach because projector will reset the GC percent to configured
value even when a single spike in RSS is observed (i.e. a spike
with magnitude >2% of memTotal)

Change-Id: I5b52f5d2b232452fdf1e54a0c0dfa9449a48ed66