MB-45006 Increase dataport bufferSize and reduce bufferTimeout
Projector flushes the data to indexer when the number of
mutations exceed bufferSize. All the mutations are queued
in a slice and they are flushed to indexer. So, if the
bufferSize is 10, there will be 10 mutations in one flush.

When indexer receives this slice of mutations, it would
distribute these mutations to stream readers. Stream
reader would then queue it in the atomic mutation queue
after updating the book-keeping. When the dataport
bufferSize is 100, it means there can be around 3 mutations
for stream worker to process before reading the next set
of mutations (assuming uniform distribution of mutations across
all the workers). However, when the dataport buffer size is 10,
it means only 10 workers will be processing the mutations while
the other 22 workers will have to wait for the mutation to get
enqueued. This is degrading the performance with high KV ops/sec
and initial index build.

To avoid this, projector now starts to flush every 100 mutations
or 1ms timeout. This makes sure that projector maintains 1ms latency
for slow mutation rates and also handles high KV ops/sec effectively.

Also, lazy trace logging is now used for endpoint buffers as a call
to Tracef can involve copying of all arguments to stack before
evaluating the log level and this can be performance sensitive in
datapath code

Change-Id: I2e579dc0c113faa2a6dd3fa4555353e54712a294