MB-44728 Part 3: Fix rebalance waitForIndexBuild hang when b/s/c dropped
This should be the final part of this MB fix. (It is independent of
and can be delivered without Part 2.)

This part prevents rebalance from waiting forever in
Rebalancer.waitForIndexBuild() if the bucket, scope, or collection of
the index got dropped during the build by moving the associated
transfer token of indexes whose metadata is not found forward to
TransferTokenCommit state instead of failing the rebalance. Indexer
will respond to b/s/c drop by deleting the metadata of indexes whose
b/s/c no longer exists (done by the recovery triggered in
Indexer.sendStreamUpdateForBuildIndex() for this case, which also logs
a Keyspace Not Found" message).

This part also contains unrelated performance improvements in
indexer/util.go GetCurrentKVTs() that skip getting the time taken and
building the verbose log message that reports it if verbose logging is
not enabled. In vitro experiements on my Mac indicate this will save
about 1us per call plus the garbage that would have been generated.

Change-Id: Ia42718aeba6f37b1cf1b4ff12d13f627b5f8bf0c