MB-44728 Part 4: Missing b/s/c error detection in user-visible errors
lifecycle.go has some cases where the missing bucket/scope/collection
canonical error string is embedded in a more detailed user-visible err
message, so Rebalancer was not detecting these as it was looking
for an exact error string match (rebalancer.go isMissingBSC()). We
want to keep the detail for users, so this fix standardizes these
cases to prepend the more detailed message with the canonical error
string inside brackets, and updates isMissingBSC() to check for both
exact matches and for the target error contained in brackets in a
longer errMsg.

Change-Id: I2df2985154bfacbf872de43a11a2e960f95bdf0f