MB-45666 Fix boot race of getClientStats and updateIndexStats2NoLock
Client bootstrap calls getClientStats() once to get a complete
IndexStats2 object to initialize local cache with. Meanwhile the
normal event-handling loop is started, which receives and processes
incoming IndexStats2 objects from periodic Lifecycle broadcasts in
updateIndexStats2NoLock(). This results in a race where a the former
could overwrite the cache with an older version than the latter has
already cached, causing later nil pointer deref to cached entry if a
bucket was added in the newer version it overwrote that the older
version does not have an entry for.

The fix is to change getClientStats() to do compare-and-swap and only
put its entry into the cache if the cache is still nil. I also removed
some cache validation code that was not needed, because either the
cache (w.clientStats) is nil (only until initialized), or it contains
a complete IndexStats2 object, as we never cache incomplete objects,
so if the cache pointer is non-nil we do not have to run through it
to see if it is complete or not.

I also made a couple performance improvements:
1. Eliminated the unnecessary cloning of the indexStats2 when it
   gets cached.
2. 14 cases: skip converting byte slices to strings for Debug log
   messags if they won't be logged.

Finally, I changed the Indexer broadcast to send a complete object
every 5 minutes in case the original response to getClientStats() got
lost or corrupted. Otherwise the client would have no list of indexes
indefinitely (until either an index or a bucket was created or
dropped).

Change-Id: Ib807a653770b0c21acf9353c6917ea62e1ac576b