MB-45743 Part 1: Fix stats div by zero and receiver performance
plasmaSlice.handleN1QLStorageStatistics() could divide by zero when
computing avg_item_size and avg_page_size stats, as after the checks
for non-zero were done it reloaded the denominators, which thus
could have changed since the checks. Fixed to load the denominators
into local variables, check those for non-zero, and then use the
local vars for the division operations.

values.go: Twelve low-level stats read functions were using value
receivers, causing an unneeded copy to be made every time a stats
value was read. Changed them to pointer receivers to avoid the extra
CPU use and garbage generation.

Initially I thought the values.go value receivers might be the cause
of the current MB, but as far as I can determine they are not, as the
objects being copied contain pointers to the stats values, not the
values themselves, and these pointers apparently never change, so the
atomic functions will be synchronizing against the same memory
location whether the receiver object is copied or not.

Change-Id: If3bb9c079651ad5533c8b24c9bb8f82322a0b1fa