MB-45919 Part 1: Add timing and logging of slow indexer msg processing
This patch adds tracking the time it takes to process each
indexer message and logs info about this at:
  1. Warn level for all slow messages, whatever their types
  2. Info level for non-slow admin (DDL) messages
  3. Debug level for non-slow non-admin messages
Slow is defined as > 1 minute to process.

The vast majority of messages fall into category 3 so will not be
logged unless Debug or higher logging level is enabled. This avoids
flooding the indexer.log.

This new logging will enable us to more easily troubleshoot cases
where the indexer becomes unresponsive.

Change-Id: Ibc375bbd46016db0a9936a1fbf3a82a5f693af6d