MB-45461 Give equal weightage to index movement and resource variation
If the system resource cost is given more weightage, planner can move
more indexes to being more evenness in the cluster. Similarly, if the
data and index movement cost is higher, planner will move less data
but the index load distribution in the cluster will be skewed. Hence,
giving equal weightage to load distribution in the cluster and index,
data movements in the cluster

Change-Id: I10d7fd4ff1ec314cc6ccf92140253827ba4f2974