MB-45461 Reduce IterationPerTemp and change initial temperature calculation
a. The current default IterationPerTemp of 1000 can move upto 1000 indexes
in one iteration and this can lead to large number of index movements in
the cluster. Reducing it to 100 to reduce the overall index movement

b. The initial temperature is calculated based on the difference in
initial cost and the variation threshold. This should help planner to
find better solution with less iterations. Otherwise, planner would
have to run for more number of iterations and this can lead to lower
quality solution

Change-Id: I9642e2a219592be58d902c66cd0d3b55ef2d95d0