MB-45553 Part 1: Split waitersMu mutex out of generic mu mutex
This patch partially fixes the MB by changing ServiceMgr.waiters field
from sharing the main mu mutex to having its own waitersMu mutex. This
will enable indexer to respond to a GetTaskList or GetTopologyChange
cancel (from ns_server via cbauth) immediately.

A follow-up fix is still needed to address the problem of the initial
API processing needing the mu mutex to copy and check the current
state object. The mutex is needed even just to check the current
state's revision number, as well as to copy and return current state
if caller passed a nil revision or current revision is newer.

Change-Id: I5a23f724ad61f7519efbdf25f93293423f67ca4a