MB-45553 Part 2: Split stateMu mutex out of generic mu mutex
This patch completes the fix of this MB by changing ServiceMgr.state
field from sharing the main mu mutex to having its own stateMu mutex.
This enables indexer to respond to a GetTaskList or GetTopologyChange
call (from ns_server via cbauth) immediately. (Part 1 enabled it to
respond to a cancel immediately, but not to new calls.)

With Part 2, the mu mutex is not needed at all for GetTaskList or
GetTopologyChange, and it is held for significantly narrower ranges
of code elsewhere. In particular it is not gotten anymore in the main
PrepareTopologyChange or StartTopologyChange bodies (but is still
needed by some of their children)

Change-Id: I3254dcb4aa1d27ea87fb3ffe6934a459d22cf533