MB-45998 Skip StreamEnd from previous sessions at dataport
If Indexer closes/opens projector stream for the same keyspace in quick
succession, it is possible for StreamEnd from previous session to arrive
at dataport after StreamBegin from the new session. This can mess up
dataport bookkeeping and lead to filtering of mutations. StreamEnd
should be allowed to cleanup dataport bookkeeping if it belongs to the
same sessionId(opaque) as StreamBegin.

Change-Id: Iac6c152619201b911919a4a674d4574ea3e71b13