MB-46054 Handle UNKNOWN_COLLECTION status at timekeeper
Clean-up the keyspace when indexer receives UNKNOWN_SCOPE
or UNKNOWN_COLLECTION status with StreamBegins. Indexer does
validate the collection with memcached (using GetCurrentKVTs
in handleBuildIndex) and with ns_server (using validateKeyspace
in sendStreamUpdateForBuildIndex).

Inspite of both this checks being successful, if stream begin
returns STREAM_UNKNOWN_COLLECTION, then it is only because the
collection has been deleted. Hence, clean-up the entire keyspace
in such a case

Change-Id: Ie1d8530721263c0b8d4c10b56b5a3a9b3e88ba90