MB-46211: Fix indexes lost on upgrade in mixed mode cluster.
In mixed mode cluster the collection Id from cinfo and metadata repo
are different. If bucket is flushed before offline upgrade this is
resulting into nil snapshot. On bootstrap after upgrade recovered indexes
stream to be set to NIL_STREAM and state is set INDEX_STATE_CREATED.
This is causing Index manager to delete the indexes recovered.

Change-Id: I0b582e76636eacff989cf1b47254fec5f06d1617