MB-44158 : Skip fetching or storing stats in MetadataProvider
stats are not used in DDL Service Manager. Waiting for
OPCODE_CLIENT_STATS response when starting a watcher is leading to
timeouts in newMetadataProvider. As this op is not async and is
dispathed with less priority, it should wait for bootstap and msgs in
expedites channel. Can try making this op Async in another patch. This
wait time for stats can anyways be avoided when not needed.

Change-Id: I3dad753357cdd68fa394a8b3822d5c975da34694