MB-46274 Decrement pendingBuild counter if index is dropped during rebalance
The "pendingBuild" counter at rebalancer keeps a track of the number
of on-going index builds on the node during rebalance. This counter
prevents rebalancer from dropping indexes when there are any pending
builds. This counter increments when build starts and decrements when
build finishes.

However, if bucket/scope/collection is dropped while build is in
progress, this counter is not decremented. This caused rebalancer to
think that there are some on-going builds and therefore it would
wait forever for dropping the indexes.

This patch fixes the issue by decrementing the counter when rebalancer
see that the index's bucket/scope/collection are dropped during
rebalance

Change-Id: Icd748eb4d98fe3e57e5718788bb3750b228b4c15