MB-46219: Fix planner to move more than 100 indexes in node removal case
The planner can move IterationPerTemp number of indexes before it
checks for variance. If variance is large enough, the planner
will choose not to move any more indexes believing that it
has converged. But if there are any deleted indexer nodes,
the planner may not have converged as there might be more
then IterationPerTemp number of index movements required.

The solution proposed here forces planner to run until all
the indexes from the deleted indexer nodes get moved. But the
number of iterations are still capped at a large value.

If there are replicas and partitions to be repaired, non deleted
nodes can become constrined and the more number of index
movements will be required. Hence, maximum number of iterations
per temperature still may not be enough. But the after the
third planner iteration (attempt), the replica and partition
repair gets disabled.

If there are existing constrained nodes in the cluster due
to resource constrints, more number of index movements will be
required and maximum number of iterations per temperature still
may not be enough. But after the sixth planner iteration (attempt),
the resource constrants will be relaxed.

This ensures that the planner will converge if there are some
deleted nodes in the cluster - without loosing any indexes.

Change-Id: Iecf12df1219a73887a44b86b83107edbf7379292