MB-46466 Prioritise abortCh over endpoint dataCh to prevent out of order seqnos
When a VBucketWorker is asked to exit, it does the following:

a. Close the worker's runFinCh and wait for run() to exit
b. The worker::run() is a go-routine that listens to mutations
   from KVData and runFinCh
c. When run() selects the closed runFinCh during execution, this
   method would exist
d. However, in cases where there are mutations in dataCh and
   runFinCh is closed, golang is free to select either of the
   channels - the behaviour is random here
e. If golang selects 2 mutations on dataCh with first one being
   a snapshot message and second one being any mutation with a
   seqno, then these mesasges would be pushed to endpoint using
   failSafeOpAsync2
f. The same problem exists in failSafeOpAsync2 as well - golang
   is free to select the abortCh (which is nothing but the closed
   runFinCh) or the endpoint dataCh
g. If golang selects abortCh for snapshot message and endpoint's
   dataCh for the next message, endpoint can report a violation

To fix this, the failSafeOpAsync2 code is updated to prioritise
the abortCh over endpoint's dataCh. This will make sure that
events will not be pushed out of order. Also, additional logging
is introduced if flushing the snapshot message to endponit is
aborted

Change-Id: I6b3baecef26d917ca67bf17c6877fc3156eaf4c9