MB-46575 : Update ServiceMap and retry if cluster version is old.
If we are seeing a non default collection means that KV is allowing to
create collection but we did not get the updated cluster version yet so
if we are seeing a Create Index on a named collection and cluster
version is still old might be due to stale serviceMap. So
refreshServiceMap and updateIndexerList and retry once.

Change-Id: I377db0cf5a0c817766482e34002a8b80db590759