MB-46513: Indexes did not get cleaned up after quorum_failover followed by dropping all collections
Root cause here is bucket flush and dropCollection events coinsided and caused dropCollection event to be missed.
The fix addresses missed dropCollection event by validating collection keyspace
during open_stream for MAINT_STREAM and calls processDropCollection if collection is not found.
This fix should reduce the probablity of issue, it is a required but not a sufficient fix.
To fully fix the issue we will additionally need to improve MonitorKeyspace to process non-defered index deletion,
which is a bigger change and will be addressed seperately.

Change-Id: I6bf87ec02d13bb32e60071607b34af7621979623