MB-46582 Increase the size of smallSnapshotThreshold for INIT_STREAM
When stream reader sends HWT to timekeeper, the seqno. of a vbucket
need not be aligned to the snapshot boundaries. E.g, If a snapshot
boundary of a vbucket is [10, 20], then seqno can be anywhere between
10 and 20.

When timekeeper receives such a HWT and tries to generate a
stability timestamp, it will try to align the seqno to the
last snapshot boundary to keep the indexer snapshotting
consistent with KV snapshots. In the above example, timekeeper
would create the timestamp as seqno: 10, snapshot: [5, 10] if
the previous snapshot is [5, 10].

This alignment to snapshot boundary happens only if the snapshot
is small i.e. if the snapshot range is <30 seqnos. This is being
governed by the setting "indexer.settings.smallSnapshotThreshold"

In some cases, the snapshot sent by KV are larger than 30 seqnos
(especially for INIT_STREAM) and this is leading to non-aligned
flushes. As merge to MAINT_STREAM can not happen if the timestamp
is not snapshot aligned, index merge is getting delayed.

This fix addresses the issue by increasing the value of
smallSnapshotThreshold for INIT_STREAM. With collections, it is
possible that seqnos of a collection are spaced out father and
KV can generate larger snapshots.

Change-Id: I9935e2271b8a528c3b36e44ba762f2be1e9f5298