MB-46376 Eliminate panics, add syncs to memdb snapshot persistence
memdb.go Visitor() method had explicit panics in three places, none of
which needed to be panics. This happened in MB-45743 (of which this MB
is an extension) where the second of these got hit by the call sequence

1. memdb_slice_impl.go OpenSnapshot() - calls Open() on s, which is
   the type memdbSnapshot grandparent of the snapshot, incrementing
   the grandparent's refCount but not the actual snapshot's refCount.
   The actual snapshot is s.info.MainSnap (type memdb.Snapshot) and
   has its own separate refCount field which is not incremented here.
2. memdb_slice_impl.go doPersistSnapshot()
3. memdb.go StoreToDisk()
4. memdb.go Visitor() - panics when NewIterator() returns nil
5. memdb/iterator.go NewIterator() - returns nil because snap.Open()
   returned false.
6. memdb.go Snapshot.Open() - returns false because it sees the actual
   snapshot's refCount == 0, meaning the snapshot is already closed
   and possibly garbage collected.

The case triggering this was already fixed in MB-45743, but we should
not use panic to report errors, so the current MB replaces these with
normal errors, which will cause persistence to be skipped for a closed
snapshot.

The snapshot persistence code was also reporting success without waiting
for the bytes to reach disk. This is also fixed here by adding os.Sync()
calls. (bufio.Writer.Flush() does not wait for disk.)

Change-Id: Iace2276bb97322d3ce8a61cc829f7944b3468842