MB-46649 prevent snapshot clone for deleted index
When an index gets deleted, storage manager
a. destroys index snapshot
b. deletes its entry from indexSnapMap
c. notifies scan coordinator of deletion which also destroys its own
copy of snapshot

Both (a) and (c) decrement the snapshot refCount, after which any
Close() call on plasma slice would see refCount=0.

However, it is possible that concurrently a request to get the latest
snapshot was sent from scan coordinator to storage manager for a
stale=false scan. This can cause listenSnapshotReqs to atomically read
the indexSnapMap before (b) above. This method can access snapC.snap
after (c) and again call CloneIndexSnapshot on the snapshot and send it
to scan coordinator for processing. This can lead to index scan to start
on a slice which has started processing Close() as refCount was 0.

Change-Id: Iafe191dab141bcac67a7e3f22b5c56ed04d24942