MB-46323 Move indexer and projector go-runtime to version 1.16.5
Go 1.13 has madv_free set to true (and madvdontneed=false). Because
of this, under certain conditions, go runtime is not returning the
memory back to OS. This is leading to increas in RSS and plasma is
reducing it's memory quota.

In Go 1.16, this change has been reverted i.e. madvdontneed is set
to true. Also, some other issues that were observed during CC
testing are fixed with Go 1.16

Change-Id: Ia04a21b3506388f9c9d218ce71af3f4989abc58b