MB-37586 Add num_disk_snapshots stat (number that currently exist)
This MB adds a new num_disk_snapshots stat that tracks the number of
disk snapshots that currently exist for the slice. This complements the
prior stats num_commits (number of disk snapshots ever written for
the slice) and num_snapshots (number of disk + memory snapshots ever
created for the slice).

num_disk_snapshots is set in:
  o memdb  - cleanupOldSnapshotFiles()
  o plasma - cleanupOldRecoveryPoints()

These are called from doPersistSnapshot() in both memdb and plasma after
creating the new disk snapshot, to delete excess old disk snapshots, so
will get the new correct number of disk snapshots that exist after both
creation and deletion of disk snapshots.

Change-Id: I5e00b68f3203a2af6aa9396235ddfab0739cd4e2