MB-47047: [Backport to 7.0.1]-Internal Server error is raised while performing backup on a index node using cbbackupmgr
Added Url encoding for bucket scope and collection name, which was missing at some places while making a call to /getLocalIndexMetadata url.
This was causing FormParsing to silently skip these values if there was a special character such as % in the bucket, scope or collection name.

Backport to 7.0.1 from 7.1.0 of MB-46895

Change-Id: If59667fc35f932bcd082491552d1b4e6e20cf780