MB-46532 Fix one Planner performance issue (CPU)
Backport one of two fixes from MB-43285 to 6.6.3:

planner/proxy.go processDeleteToken() and processDropInstanceToken()
will build token lookup maps so the loop iterating over tokens will
now precede, instead of contain, the two nested loops iterating over
indexers and indexes. This will save CPU and a lot of unneeded
garbage generation when there are many tokens.

The other fix from the original MB is not backportable because it
requires new REST APIs, which means the code must be able to detect
whether any nodes are pre-6.6.3 and use the old APIs and algorithm
instead, however the cluster info we have includes the major and minor
version numbers only, not the fix release number, i.e. we can tell
6.6 from 6.5 but not 6.6.3 from 6.6.2.

Not backporting this change is not a significant loss because it is a
very minor performance enhancement that gets delete and drop instance
token paths from metakv instead of the full tokens. There was a later,
more powerful fix (MB-44269) that was already backported to 6.6.3 (via
MB-44408) that fixed token lookups in metakv to get all the tokens out
of the single large response generated by the metakv list call, whereas
previously GSI code mistakenly thought the tokens were not included in
that response and so proceeded to retrieve them from metakv one by one.
That backport superceded much of the change in ddl_service_manager.go
and gives a bigger performance boost than the unbackportable change.

The result of not backporting this second change is that token.go and
ddl_service_manager.go have no code changes to backport -- just a few
comment additions to existing code. There were also a lot of logging
updates in proxy.go that cannot be backported because the functions they
are in do not exist in 6.6.3 (all those with names of the form
"fooResp"), and likewise the mass deletions of the functions those
replaced in 7.0.0 (same names minus the "Resp" suffix) could not be
backported because they are the versions still in use in 6.6.3.

Change-Id: I7a6e417aaaed2158ccf581e8427dd8c493dff00a