MB-47323 Stop mutation scatter before draining the mutch in kvdata
BP of MB-43259 to 6.6.3

When feed decides to clean-up the bucket, it spawns a go-routine
to drain the mutation channel in kvdata. This is done to ensure that
the StreamEnd messages published by dcp_feed.go routines will be
consumed if the kvdata mutch is full.

When this go-routine is spawned in cleanupBucket, there will be
two consumers of kvdata's mutch. In such a case, it can be possible
that important control messages like Snapshot marker message gets
drained while the mutations belonging to that snapshot are processed
at mutch in kvdata. This breaks the DCP logic as indexer always
expects a snapshot marker message before a mutation from that snapshot.

The issue is fixed in this patch by stopping the runScatter method
before processing any mutations.

Change-Id: I30bb3c2a9c353474b2893154b125e2c6747e20e3